requirements.txt
Flask==3.1.0
ldap3==2.9.1
python-dotenv==1.0.1
